Moved AfrikaExceptions into separate package
